Switched intercom to layout
