+++ frontend/src/components/utilities/config/index.ts
@@ -4,11 +4,12 @@ const POSTHOG_HOST =
   process.env.NEXT_PUBLIC_POSTHOG_HOST! || "https://app.posthog.com";
 const STRIPE_PRODUCT_PRO = process.env.NEXT_PUBLIC_STRIPE_PRODUCT_PRO!;
 const STRIPE_PRODUCT_STARTER = process.env.NEXT_PUBLIC_STRIPE_PRODUCT_STARTER!;
+const INTERCOM_ID = process.env.NEXT_PUBLIC_INTERCOM_ID!;
 
 export {
   ENV,
+  INTERCOM_ID,
   POSTHOG_API_KEY,
   POSTHOG_HOST,
   STRIPE_PRODUCT_PRO,
-  STRIPE_PRODUCT_STARTER
-};
+  STRIPE_PRODUCT_STARTER};
